pyproject.toml
[project]
name = "brain"
version = "0.1.0"
description = "MNEMOS - Personal Cognitive Engine / Digital Virtual Brain"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "emoji>=2.15.0",
    "faiss-cpu>=1.13.0",
    "fastapi>=0.128.8",
    "httpx>=0.28.0",
    "iso8601>=2.1.0",
    "loguru>=0.7.3",
    "mkdocs-material>=9.7.6",
    "mkdocs-static-i18n>=1.3.1",
    "numpy>=2.0.2",
    "peft>=0.17.1",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.13.0",
    "python-multipart>=0.0.20",
    "python-telegram-bot[job-queue]>=22.5",
    "rich>=15.0.0",
    "sentence-transformers>=5.1.2",
    "transformers>=4.57.6",
    "trl>=0.24.0",
    "typer>=0.23.2",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=8.4.2",
    "ruff>=0.15.10",
]

# Optional LLM backends — install only what you need:
#   uv add anthropic          → Anthropic Claude API
#   uv add openai             → OpenAI / compatible API
# Ollama uses httpx (already a core dep); no extra package needed.

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
